+++ R/is_model_supported.R
@@ -47,6 +47,7 @@ supported_models <- function() {
     "BBreg",
     "betareg",
     "BFBayesFactor",
+    "bife",
     "biglm",
     "bigglm",
     "blavaan",