+++ NEWS.md
@@ -4,6 +4,10 @@
 
 * `model_info()` now only returns `TRUE` for `$is_ordinal`, when model is an ordinal or cumulative link model. In past version, `$is_ordinal` was also `TRUE` for multinomial models.
 
+## New supported model classes
+
+* `bife` (*bife*)
+ 
 ## General
 
 * Improved support for `clogit`-models.